# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobify API",
        "description": "Documentación de la API de Jobify",
        "contact": {
            "email": "soporte@jobify.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/aplicaciones": {
            "get": {
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Listar aplicaciones",
                "operationId": "a7e68a8cd95539545c8d4bdf302d0ab2",
                "responses": {
                    "200": {
                        "description": "Lista de aplicaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Aplicar a una oferta",
                "operationId": "1ed0992fb6c5a5f0dc68f6c1cf42d978",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAplicacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aplicación creada"
                    }
                }
            }
        },
        "/api/aplicaciones/{aplicacion}": {
            "put": {
                "tags": [
                    "Aplicaciones"
                ],
                "summary": "Actualizar estado de aplicación",
                "operationId": "463261645ebe1db7f2f2a48bb980edb4",
                "parameters": [
                    {
                        "name": "aplicacion",
                        "in": "path",
                        "description": "ID de la aplicación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAplicacionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aplicación actualizada"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "operationId": "4d730c449df9d7cd37cc97b2daddf8ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "422": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "ee3fc9fe808bc810a35bd43d96dada63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente"
                    }
                }
            }
        },
        "/api/candidatos": {
            "get": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Listar candidatos",
                "operationId": "1879147fe9750cfb75f91712c8ce3557",
                "responses": {
                    "200": {
                        "description": "Lista de candidatos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Crear perfil de candidato",
                "operationId": "0a035d99352c84d4805769e3b7b9b5a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCandidatoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidato creado"
                    }
                }
            }
        },
        "/api/candidatos/{candidato}": {
            "get": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Ver candidato",
                "operationId": "683a92e287219d50b79e96a779d902c1",
                "parameters": [
                    {
                        "name": "candidato",
                        "in": "path",
                        "description": "ID del candidato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del candidato"
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidatos"
                ],
                "summary": "Actualizar candidato",
                "operationId": "71d7cd9d437ebde1ada5ff26646f115c",
                "parameters": [
                    {
                        "name": "candidato",
                        "in": "path",
                        "description": "ID del candidato",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCandidatoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidato actualizado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login Request",
                "description": "Datos para iniciar sesión",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "example": "admin@jobify.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Datos para registrar un nuevo usuario",
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "title": "Nombre",
                        "description": "Nombre completo del usuario",
                        "type": "string",
                        "example": "Juan Perez"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "example": "juan@example.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "title": "Confirmación de Password",
                        "description": "Confirmación de la contraseña",
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "StoreAplicacionRequest": {
                "title": "Store Aplicacion Request",
                "description": "Datos para crear una nueva aplicación",
                "required": [
                    "trabajo_id"
                ],
                "properties": {
                    "trabajo_id": {
                        "description": "ID del trabajo al que se aplica",
                        "type": "integer",
                        "example": 1
                    },
                    "mensaje": {
                        "description": "Mensaje opcional para la aplicación",
                        "type": "string",
                        "example": "Estoy muy interesado en este puesto."
                    }
                },
                "type": "object"
            },
...
            "UpdateAplicacionRequest": {
                "title": "Update Aplicacion Request",
                "description": "Datos para actualizar el estado de una aplicación",
                "required": [
                    "estado"
                ],
                "properties": {
                    "estado": {
                        "description": "Nuevo estado de la aplicación",
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "aceptado",
                            "rechazado"
                        ],
                        "example": "aceptado"
                    }
                },
                "type": "object"
            },
            "UpdateCandidatoRequest": {
                "title": "Update Candidato Request",
                "description": "Datos para actualizar el perfil del candidato",
                "properties": {
                    "apellidos": {
                        "description": "Apellidos del candidato",
                        "type": "string",
                        "example": "Doe Smith"
                    },
                    "fecha_nacimiento": {
                        "description": "Fecha de nacimiento",
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "descripcion": {
                        "description": "Breve descripción profesional",
                        "type": "string",
                        "example": "Desarrollador apasionado..."
                    },
                    "ubicacion": {
                        "description": "Ubicación actual",
                        "type": "string",
                        "example": "Barcelona"
                    },
                    "url_cv": {
                        "description": "URL al CV (PDF, LinkedIn, etc)",
                        "type": "string",
                        "example": "https://linkedin.com/in/johndoe"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}